refactor: move page helpers to separate modules
